LayoutTests/imported/w3c/web-platform-tests/XMLHttpRequest/xmlhttprequest-sync-not-hang-scriptloader-expected.txt

CONSOLE MESSAGE: line 14: XMLHttpRequest cannot load data:,. Cross origin requests are only supported for HTTP.
CONSOLE MESSAGE: line 14: NetworkError (DOM Exception 19):  A network error occurred.


PASS Ensure that an async script added during a defer script that then does a
  sync XHR still runs 
